Add setup.py file
Have added a `setup.py` file so that users of this application are able
to easily install any dependencies locally. This requires the `pip-tools`
package. Running `pip-compile` will generate the `requirements.txt`
file based on the configuration in `setup.py` and it should be possible
to install dependencies with `pip install -r requirements.txt`