Add swagger documentation
Have added support for swagger documentation. This has meant reordering
the urlpatterns list and and adding the drf-yasg library. The library
appears to have a [bug](https://github.com/axnsan12/drf-yasg/issues/510)
in it and so only supports JSON response types. Its support for endpoints
that aren't associated with a model is lacking as this requires the use of
a custom decorator.

Have also replaced some string values with environment variables so that
these are easier to change the future. This commit should also resolve a
bug in the routes that redirected users to homepage when a 404 should
have been displayed.