Added nested endpoints to /users
A new router has been created to handle the nested endpoints under
/users. It should now be possible to filter the different surveys and
survey responses by user with nested resources. The endpoints created
are:

* /users/:user_id/surveys
* /users/:user_id/survey-responses

Note, it is still possible to create surveys and survey responses under
the wrong user ID as the POST endpoint uses the field in the request
body rather than the URL. As this field is mandatory, I will assume that
any users creating records in this way will know what they're doing.