Fix bug in db-clean command
The IDs in django tables use serial columns rather than identity
columns. This is a little old fashioned in my opinion but does keep
compatibility with older versions of PostgreSQL. As such, I've had to
re-write the `make db-clean` command so that it will correctly truncate
rows with serial primary keys.