Fix bug in db-clean and refactor SQL
The db-clean command didn't work because the wrong database credentials
were used to connect to PostgreSQL (it should use the application
database name and owner instead, unlike the db-create and db-drop).

I have also cleaned up the SQL in all of the the database operations so
that strings are safely formatted before they're executed. Typically
this would protect against SQL injections and is just best practice.